modified:   app/Http/Controllers/librodiario/LibrodiarioController.php 	modified:   app/Http/Controllers/libromayor/LibromayorController.php 	modified:   resources/views/librodiario/librodiario.blade.php 	modified:   resources/views/libromayor/libromayor.blade.php

+++ app/Http/Controllers/libromayor/LibromayorController.php
@@ -35,62 +35,64 @@ class LibromayorController extends Controller
 			foreach ($permisos as $key) {
 				$consultar = $key['PER_CONSULTAR'];
 			}
-
-			// if ($consultar == '1'){
-			// 	return 'si';
-			// }else{
-			// 	return 'no';
-			// }
-
 		} catch (\Throwable $th) {
-			return 'Error Libro Mayor 46';
+			return 'Error Libro Mayor 39';
 		}
 
 
-		try {
-
-			$clasificacion = http::withToken(Cache::get('token'))->get($this->url . '/clasificacion');
-			$clasificacionArr = $clasificacion->json();
-			//nombre de cuenta
-			$nombrecuenta = http::withToken(Cache::get('token'))->get($this->url . '/cuentas');
-			$nombrecuentaArr = $nombrecuenta->json();
-
-			$periodo = http::withToken(Cache::get('token'))->get($this->url . '/periodo');
-			$periodoArr = $periodo->json();
+		if (isset( $consultar) == '1') {
 
 
-			# code...
-		} catch (\Throwable $th) {
-			return 'error libro diario 29';
+			try {
+
+				$clasificacion = http::withToken(Cache::get('token'))->get($this->url . '/clasificacion');
+				$clasificacionArr = $clasificacion->json();
+				//nombre de cuenta
+				$nombrecuenta = http::withToken(Cache::get('token'))->get($this->url . '/cuentas');
+				$nombrecuentaArr = $nombrecuenta->json();
+
+				$periodo = http::withToken(Cache::get('token'))->get($this->url . '/periodo');
+				$periodoArr = $periodo->json();
+
+
+				# code...
+			} catch (\Throwable $th) {
+				return 'error libro Mayor 60';
+			}
+
+
+			try {
+
+				$libromayor = http::withToken(Cache::get('token'))->get($this->url . '/libromayor');
+
+				$personArr = $libromayor->json();
+
+
+				# code...
+			} catch (\Throwable $e) {
+				return 'error libro mayor 73';
+			}
+
+
+
+			try {
+				$bitacora = Http::withToken(Cache::get('token'))->post($this->url . '/seguridad/bitacora/insertar', [
+
+					"USR" => Cache::get('user'),
+					"ACCION" => 'PANTALLA METODO GET',
+					"DES" => Cache::get('user') . 'INGRESO A LA PANTALLA DE LIBRO MAYOR',
+					"OBJETO" => 'LIBROMAYOR'
+
+				]);
+			} catch (\Throwable $th) {
+				return 'Error Libro Mayor 88';
+			}
+		} else {
+			return view('Auth.no-auth');
 		}
 
 
-		try {
 
-			$libromayor = http::withToken(Cache::get('token'))->get($this->url . '/libromayor');
-
-			$personArr = $libromayor->json();
-
-
-			# code...
-		} catch (\Throwable $e) {
-			return 'error libro mayor 30';
-		}
-
-
-
-		try {
-			$bitacora = Http::withToken(Cache::get('token'))->post($this->url . '/seguridad/bitacora/insertar', [
-
-				"USR" => Cache::get('user'),
-				"ACCION" => 'PANTALLA METODO GET',
-				"DES" => Cache::get('user') . 'INGRESO A LA PANTALLA DE LIBRO MAYOR',
-				"OBJETO" => 'LIBROMAYOR'
-
-			]);
-		} catch (\Throwable $th) {
-			return 'Error Libro Mayor 92';
-		}
 
 		return view('libromayor.libromayor', compact('personArr', 'clasificacionArr', 'nombrecuentaArr', 'periodoArr'));
 	}
@@ -99,10 +101,34 @@ class LibromayorController extends Controller
 	// INSERTAR FUNCIONAL
 	public function insertar(Request $request)
 	{
-		// return $request;
+
+
+
+
+		/**
+		 * Seguridad de roles y perimisos metodo insert
+		 */
+
 		try {
 
-		
+			$search = Http::withToken(Cache::get('token'))->post($this->url . '/permisos/sel_per_obj', [
+				"PV_ROL" => Cache::get('rol'),
+				"PV_OBJ" => "LIBROMAYOR"
+			]);
+
+			$permisos = $search->json();
+			foreach ($permisos as $key) {
+				$insercion = $key['PER_INSERCION'];
+			}
+		} catch (\Throwable $th) {
+			return 'Error Libro Mayor 124';
+		}
+
+		if (isset( $insercion) == '1') {
+			// return $request;
+			try {
+
+
 				# code...
 
 				$cargo = Http::withToken(Cache::get('token'))->post($this->url . '/libromayor/insertar', [
@@ -115,7 +141,7 @@ class LibromayorController extends Controller
 
 
 				]);
-		
+
 				$abono = Http::withToken(Cache::get('token'))->post($this->url . '/libromayor/insertar', [
 
 					"COD_PERIODO" => $request->periodo,
@@ -125,38 +151,54 @@ class LibromayorController extends Controller
 
 
 				]);
-			
-		} catch (\Throwable $e) {
-			return 'Error libromayor 44';
+			} catch (\Throwable $e) {
+				return 'Error libromayor 155';
+			}
+
+
+
+
+			try {
+				$bitacora = Http::withToken(Cache::get('token'))->post($this->url . '/seguridad/bitacora/insertar', [
+
+					"USR" => Cache::get('user'),
+					"ACCION" => 'PANTALLA METODO POST',
+					"DES" => Cache::get('user') . 'INSERTO EL DATO DE ' . $request->libromayor . 'EN LA PANTALLA DE LIBRO MAYOR',
+					"OBJETO" => 'LIBROMAYOR'
+
+				]);
+			} catch (\Throwable $th) {
+				return 'Error Libro Mayor 171';
+			}
+
+			Session::flash('insertado', '1');
+		} else {
+
+
+
+			try {
+				$bitacora = Http::withToken(Cache::get('token'))->post($this->url . '/seguridad/bitacora/insertar', [
+
+					"USR" => Cache::get('user'),
+					"ACCION" => 'SIN PERMISO METODO POST',
+					"DES" => Cache::get('user') . 'INTENTO INSERTAR UN DATO  ' . $request->libromayor . 'EN LA PANTALLA DE LIBRO MAYOR',
+					"OBJETO" => 'LIBROMAYOR'
+
+				]);
+
+				Session::flash('sin permiso', '1');
+			} catch (\Throwable $th) {
+				return 'Error Libro Mayor 191';
+			}
 		}
-
-
-
-
-		try {
-			$bitacora = Http::withToken(Cache::get('token'))->post($this->url . '/seguridad/bitacora/insertar', [
-
-				"USR" => Cache::get('user'),
-				"ACCION" => 'PANTALLA METODO POST',
-				"DES" => Cache::get('user') . 'INSERTO EL DATO DE '.$request->libromayor.'EN LA PANTALLA DE LIBRO MAYOR',
-				"OBJETO" => 'LIBROMAYOR'
-
-			]);
-		} catch (\Throwable $th) {
-			return 'Error Libro Mayor 43';
-		}
-
 
 
 		Session::flash('insertado', '1');
 		return back();
-
-
-		
 	}
 
 
-	// MAYORIZACION FUNCIONAL
+	// MAYORIZACION FUNCIONAL OCUPA SEGURIDAD??
 	public function mayorizacion(Request $request)
 	{
 
@@ -170,8 +212,6 @@ class LibromayorController extends Controller
 		// return $request;
 		Session::flash('insertado', '1');
 		return back();
-
-
 	}
 
 
@@ -179,54 +219,100 @@ class LibromayorController extends Controller
 	public function actualizar(Request $request)
 	{
 
-		try {
-			if($request->transaccion == '1'){
-			$actualizar = Http::withToken(Cache::get('token'))->put($this->url . '/libromayor/actualizar/' . $request->f, [
 
-
-
-
-				"COD_PERIODO" => $request->periodo,
-				"NOM_CUENTA" => $request->cuenta,
-				"SAL_DEBE" => $request->saldo,
-				"SAL_HABER" => 0,
-			]);
-		}elseif ($request->transaccion == '0'){
-
-			$actualizar = Http::withToken(Cache::get('token'))->put($this->url . '/libromayor/actualizar/' . $request->f, [
-
-				"COD_PERIODO" => $request->periodo,
-				"NOM_CUENTA" => $request->cuenta,
-				"SAL_DEBE" => 0,
-				"SAL_HABER" => $request->saldo,
-			]);
-
-		}
-
-			# code...
-		} catch (\Throwable $e) {
-			# code...
-			return 'error libro mayor 76';
-		}
+		/**
+		 * Seguridad de roles y perimisos metodo update
+		 */
 
 		try {
-			$bitacora = Http::withToken(Cache::get('token'))->post($this->url . '/seguridad/bitacora/insertar', [
 
-				"USR" => Cache::get('user'),
-				"ACCION" => 'PANTALLA METODO PUT',
-				"DES" => Cache::get('user') . 'ACTUALIZO EL DATO DE  '.$request->libromayor.'EN LA PANTALLA DE LIBRO MAYOR',
-				"OBJETO" => 'LIBROMAYOR'
-
+			$search = Http::withToken(Cache::get('token'))->post($this->url . '/permisos/sel_per_obj', [
+				"PV_ROL" => Cache::get('rol'),
+				"PV_OBJ" => "LIBROMAYOR"
 			]);
+
+			$permisos = $search->json();
+			foreach ($permisos as $key) {
+				$update = $key['PER_ACTUALIZACION'];
+			}
 		} catch (\Throwable $th) {
-			return 'Error Libro Mayor 43';
+			return 'Error Libro Mayor 239';
 		}
 
+		if (isset( $update) == '1') {
+
+			try {
+				if ($request->transaccion == '1') {
+					$actualizar = Http::withToken(Cache::get('token'))->put($this->url . '/libromayor/actualizar/' . $request->f, [
 
 
 
 
-		Session::flash('actualizado', '1');
+						"COD_PERIODO" => $request->periodo,
+						"NOM_CUENTA" => $request->cuenta,
+						"SAL_DEBE" => $request->saldo,
+						"SAL_HABER" => 0,
+					]);
+				} elseif (
+					$request->transaccion == '0'
+				) {
+
+					$actualizar = Http::withToken(Cache::get('token'))->put($this->url . '/libromayor/actualizar/' . $request->f, [
+
+						"COD_PERIODO" => $request->periodo,
+						"NOM_CUENTA" => $request->cuenta,
+						"SAL_DEBE" => 0,
+						"SAL_HABER" => $request->saldo,
+					]);
+				}
+
+				# code...
+			} catch (\Throwable $e) {
+				# code...
+				return 'error libro mayor 272';
+			}
+
+			try {
+				$bitacora = Http::withToken(Cache::get('token'))->post($this->url . '/seguridad/bitacora/insertar', [
+
+					"USR" => Cache::get('user'),
+					"ACCION" => 'PANTALLA METODO PUT',
+					"DES" => Cache::get('user') . 'ACTUALIZO EL DATO DE  ' . $request->libromayor . 'EN LA PANTALLA DE LIBRO MAYOR',
+					"OBJETO" => 'LIBROMAYOR'
+
+				]);
+			} catch (\Throwable $th) {
+				return 'Error Libro Mayor 285';
+			}
+
+
+
+
+
+			Session::flash(
+				'actualizado',
+				'1'
+			);
+		} else {
+
+			try {
+				$bitacora = Http::withToken(Cache::get('token'))->post($this->url . '/seguridad/bitacora/insertar', [
+
+					"USR" => Cache::get('user'),
+					"ACCION" => 'SIN PERMISO METODO PUT',
+					"DES" => Cache::get('user') . 'INTENTO ACTUALIZAR UN DATO  ' . $request->libromayor . 'EN LA PANTALLA DE LIBRO MAYOR',
+					"OBJETO" => 'LIBROMAYOR'
+
+				]);
+
+				Session::flash(
+					'sin permiso',
+					'1'
+				);
+			} catch (\Throwable $th) {
+				return 'Error Libro Mayor 313';
+			}
+		}
 		return back();
 
 
@@ -238,26 +324,68 @@ class LibromayorController extends Controller
 	public function eliminar(Request $request)
 	{
 
-		$delete = Http::withToken(Cache::get("token"))->delete($this->url . '/libromayor/eliminar/' . $request->f,);
 
-
-
+		/**
+		 * Seguridad de roles y perimisos metodo delete
+		 */
 
 		try {
-			$bitacora = Http::withToken(Cache::get('token'))->post($this->url . '/seguridad/bitacora/insertar', [
 
-				"USR" => Cache::get('user'),
-				"ACCION" => 'ELIMINO UN DATO',
-				"DES" => Cache::get('user') . 'ELIMINO EL DATO CON CODIGO DE  '.$request->f. 'EN LA PANTALLA DE LIBRO MAYOR',
-				"OBJETO" => 'LIBROMAYOR'
-
+			$search = Http::withToken(Cache::get('token'))->post($this->url . '/permisos/sel_per_obj', [
+				"PV_ROL" => Cache::get('rol'),
+				"PV_OBJ" => "LIBROMAYOR"
 			]);
+
+			$permisos = $search->json();
+			foreach ($permisos as $key) {
+				$eliminacion = $key['PER_ELIMINACION'];
+			}
 		} catch (\Throwable $th) {
-			return 'Error Libro Mayor 43';
+			return 'Error Libro Mayor 344';
 		}
 
+		if ( isset( $eliminacion )== '1') {
 
-		Session::flash('eliminado', '1');
+			$delete = Http::withToken(Cache::get("token"))->delete($this->url . '/libromayor/eliminar/' . $request->f,);
+
+
+
+
+			try {
+				$bitacora = Http::withToken(Cache::get('token'))->post($this->url . '/seguridad/bitacora/insertar', [
+
+					"USR" => Cache::get('user'),
+					"ACCION" => 'ELIMINO UN DATO',
+					"DES" => Cache::get('user') . 'ELIMINO EL DATO CON CODIGO DE  ' . $request->f . 'EN LA PANTALLA DE LIBRO MAYOR',
+					"OBJETO" => 'LIBROMAYOR'
+
+				]);
+			} catch (\Throwable $th) {
+				return 'Error Libro Mayor 364';
+			}
+
+
+			Session::flash('eliminado', '1');
+		} else {
+
+			try {
+				$bitacora = Http::withToken(Cache::get('token'))->post($this->url . '/seguridad/bitacora/insertar', [
+
+					"USR" => Cache::get('user'),
+					"ACCION" => 'SIN PERMISO METODO DELETE',
+					"DES" => Cache::get('user') . 'INTENTO ELIMINAR UN DATO  ' . $request->f . 'EN LA PANTALLA DE LIBRO MAYOR',
+					"OBJETO" => 'LIBROMAYOR'
+
+				]);
+
+				Session::flash(
+					'sin permiso',
+					'1'
+				);
+			} catch (\Throwable $th) {
+				return 'Error Libro Mayor 386';
+			}
+		}
 		return back();
 
 
@@ -269,6 +397,7 @@ class LibromayorController extends Controller
 	// FUNCION PARA PDF FUNCIONAL 
 	public function pdf()
 	{
+
 		$libromayor = http::withToken(Cache::get('token'))->get($this->url . '/libromayor');
 
 		$mayor = $libromayor->json();
